chore: rename generate lead key
