+++ api/src/domains/teams/routes.js
@@ -11,8 +11,8 @@ const teamRoutes = () => {
     const controller = TeamController(repository)
 
     router.post('/', controller.addTeam)
-    router.post('/:id/generate-lead-keys', controller.generateLeadKeys)
-    router.post('/delete-lead-key', controller.deleteLeadershipKey)
+    router.post('/:id/generate-leadership-keys', controller.generateLeadershipKeys)
+    router.post('/delete-leadership-key', controller.deleteLeadershipKey)
     router.get('/', controller.getTeams)
     router.get('/:id', controller.getTeamById)
     router.get('/:id/shifts', controller.getShifts)
